+++ app/api/api/serializers.py
@@ -6,7 +6,7 @@ import openpyxl
 from django.contrib.auth.models import User
 from drf_dynamic_fields import DynamicFieldsMixin
 from mapping.permissions import has_editorship, is_admin, is_az_function_user
-from mapping.services.services_rules import analyse_concepts, get_mapping_rules_json
+from mapping.services.rules import analyse_concepts, get_mapping_rules_json
 from openpyxl.workbook.workbook import Workbook
 from rest_framework import serializers
 from rest_framework.exceptions import NotFound, ParseError, PermissionDenied
