+++ app/api/mapping/management/commands/generate_rules_json.py
@@ -1,7 +1,7 @@
 import json
 
 from django.core.management.base import BaseCommand
-from mapping.services.services_rules import get_mapping_rules_json, make_dag
+from mapping.services.rules import get_mapping_rules_json, make_dag
 from shared.data.models import MappingRule
